Separated the Reservation booking in a user control
