+++ HolidayDestinations/HolidayDestinations.Web/Controls/ReservationBook.ascx
@@ -1,0 +1,10 @@
+﻿<%@ Control Language="C#" AutoEventWireup="true" CodeBehind="ReservationBook.ascx.cs" Inherits="HolidayDestinations.Web.Controls.ReservationBook" %>
+
+<asp:UpdatePanel runat="server" ID="UpdatePanel">
+    <ContentTemplate>
+        <asp:Button Text="Book" ID="btnBook" OnClick="btnBook_Click" runat="server"></asp:Button>
+    </ContentTemplate>
+    <Triggers>
+        <asp:AsyncPostBackTrigger ControlID="btnBook" EventName="Click" />
+    </Triggers>
+</asp:UpdatePanel>
