+++ HolidayDestinations/HolidayDestinations.Web/DestinationDetails.aspx.cs
@@ -12,10 +12,6 @@
     {
         [Inject]
         public IDestinationsService DestinationsService { get; set; }
-        [Inject]
-        public IReservationsService ReservationsService { get; set; }
-
-
 
         protected void Page_Load(object sender, EventArgs e)
         {
@@ -33,19 +29,6 @@
                 return null;
             }
             return this.DestinationsService.GetById(idToSearch);
-        }
-
-        protected void btnBook_Click(object sender, EventArgs e)
-        {
-            if (!Page.IsValid)
-            {
-                return;
-            }
-            var currentUserId = this.Page.User.Identity.GetUserId();
-            var currentDestinationId = int.Parse(this.Request.QueryString["id"]);
-            this.ReservationsService.CreateReservation(currentUserId, currentDestinationId);
-
-            this.Response.Redirect("/Account/Reservations");
         }
     }
 }
