+++ HolidayDestinations/HolidayDestinations.Web/Web.config
@@ -22,6 +22,7 @@
       </namespaces>
       <controls>
         <add src="~/Controls/HomePagination.ascx" tagName="HomePagination" tagPrefix="holidayDestinations"/>
+        <add src="~/Controls/ReservationBook.ascx" tagName="ReservationBook" tagPrefix="holidayDestinations"/>
         <add assembly="Microsoft.AspNet.Web.Optimization.WebForms" namespace="Microsoft.AspNet.Web.Optimization.WebForms" tagPrefix="webopt" />
       </controls>
     </pages>
